PostMessageTransport. Реализован с использованием EventBus
